address more comments
