+++ plugins/plugin-chart-echarts/src/utils/series.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-underscore-dangle */
 /**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -16,55 +17,23 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { DataRecordValue } from '@superset-ui/chart';
-import { EchartsBaseTimeseriesSeries, EchartsTimeseriesDatum } from '../Timeseries/types';
+import { TimeseriesDataRecord } from '../Timeseries/types';
 
-export const extractTimeseriesSeries = (
-  data: EchartsTimeseriesDatum[],
-): Partial<echarts.EChartOption.Series>[] => {
-  const series = [] as EchartsBaseTimeseriesSeries[];
+// eslint-disable-next-line import/prefer-default-export
+export function extractTimeseriesSeries(
+  data: TimeseriesDataRecord[],
+): echarts.EChartOption.Series[] {
+  if (data.length === 0) return [];
+  const rows = data.map(datum => ({
+    ...datum,
+    __timestamp: new Date(datum.__timestamp),
+  }));
 
-  const keys = data.length > 0 ? Object.keys(data[0]).filter(key => key !== '__timestamp') : [];
-
-  const rawSeries: Record<string, [Date, DataRecordValue][]> = keys.reduce(
-    (obj, key) => ({
-      ...obj,
-      [key]: [],
-    }),
-    {},
-  );
-
-  data.forEach(row => {
-    // eslint-disable-next-line no-underscore-dangle
-    const timestamp = new Date(row.__timestamp);
-    keys.forEach(key => {
-      rawSeries[key].push([timestamp, row[key]]);
-    });
-  });
-
-  Object.entries(rawSeries).forEach(([key, value]) => {
-    series.push({
+  return Object.keys(rows[0])
+    .filter(key => key !== '__timestamp')
+    .map(key => ({
       name: key,
-      data: value,
-    });
-  });
-
-  // @ts-ignore
-  return series;
-};
-
-export const extractSeriesBase = (series: echarts.EChartOption.Series[]): number | null => {
-  let minValue: number | null = null;
-  series
-    .filter(row => row !== undefined)
-    .forEach(seriesEntry => {
-      if (seriesEntry?.data) {
-        // @ts-ignore
-        seriesEntry.data.forEach(row => {
-          minValue =
-            minValue === null || minValue === undefined || row[1] < minValue ? row[1] : minValue;
-        });
-      }
-    });
-  return minValue;
-};
+      // @ts-ignore
+      data: rows.map(datum => [datum.__timestamp, datum[key]]),
+    }));
+}
