+++ plugins/plugin-chart-echarts/test/Timeseries/buildQuery.test.ts
@@ -16,7 +16,6 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import 'babel-polyfill';
 import buildQuery from '../../src/Timeseries/buildQuery';
 
 describe('EchartsTimeseries buildQuery', () => {
