+++ plugins/plugin-chart-word-cloud/test/plugin/buildQuery.test.ts
@@ -1,4 +1,3 @@
-import 'babel-polyfill';
 import buildQuery from '../../src/plugin/buildQuery';
 
 describe('WordCloud buildQuery', () => {
